requirements.txt
altair==4.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.5
bokeh==2.4.3
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
cmdstanpy==1.0.7
colorlover==0.3.0
commonmark==0.9.1
compress-pickle==2.1.0
convertdate==2.4.0
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
ephem==4.1.3
executing==1.0.0
fonttools==4.37.1
gitdb==4.0.9
GitPython==3.1.27
hijri-converter==2.2.4
holidays==0.15
hydralit-components==1.0.10
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.2
ipython==8.5.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
multitasking==0.0.11
nest-asyncio==1.5.5
numpy==1.23.2
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.31
prophet==1.1.1
protobuf==3.20.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
PyMeeus==0.5.11
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
rich==12.5.1
semver==2.13.0
setuptools-git==1.2
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.5.0
streamlit==1.12.2
tenacity==8.0.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
widgetsnbextension==4.0.3
yfinance==0.1.74
zipp==3.8.1
